@@ -12,18 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
+    "@angular/animations": "^8.2.5",
     "@angular/cdk": "^6.4.7",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
+    "@angular/common": "^8.2.5",
+    "@angular/compiler": "^8.2.5",
+    "@angular/core": "^8.2.5",
+    "@angular/forms": "^8.2.5",
     "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/platform-server": "^7.0.0",
-    "@angular/router": "^7.0.0",
+    "@angular/platform-browser": "^8.2.5",
+    "@angular/platform-browser-dynamic": "^8.2.5",
+    "@angular/platform-server": "^8.2.5",
+    "@angular/router": "^8.2.5",
     "@auth0/angular-jwt": "^2.0.0",
     "@ngrx/store": "^6.1.0",
     "@ngrx/store-devtools": "^6.1.0",
@@ -32,16 +31,17 @@
     "ng-recaptcha": "^4.0.0",
     "rxjs": "^6.3.3",
     "rxjs-compat": "^6.3.3",
-    "zone.js": "^0.8.14"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "^7.0.2",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/language-service": "^7.0.0",
+    "@angular-devkit/build-angular": "~0.803.3",
+    "@angular/cli": "^8.3.3",
+    "@angular/compiler-cli": "^8.2.5",
+    "@angular/language-service": "^8.2.5",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.5",
     "@types/node": "^6.0.118",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -53,6 +53,6 @@
     "protractor": "^5.4.1",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "~3.1.3"
+    "typescript": "~3.5.3"
   }
 }
