@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Rx';
 import { Store } from '@ngrx/store';
@@ -13,8 +12,7 @@ export class PresentService extends BaseHttpService {
 
     otherPresents:Observable<Array<any>>;
 
-    constructor(private http:Http,
-                private store: Store<any>,
+    constructor(private store: Store<any>,
                 private httpClient: HttpClient) { 
         super();
         this.otherPresents = this.store.select("presents");
